Versão alfa

+++ src/main/java/br/com/developed/jticket/models/EstoqueDisponivel.java
@@ -15,5 +15,5 @@ import lombok.Setter;
  */
 
 public interface EstoqueDisponivel {
-    Integer getEstoque();
+    Double getEstoque();
 }
